Gdal utils raster (#3153)
* Move Raster to private space.

* Split up GDALUtils and move to private.

* Missed header.

* BaseBand.

* Move GDALUtils to private.

* Remove unnecessary file.

* Remove dead code.
