+++ filters/CropFilter.cpp
@@ -39,7 +39,7 @@
 #include <pdal/Polygon.hpp>
 #include <pdal/util/Bounds.hpp>
 #include <pdal/util/ProgramArgs.hpp>
-#include <pdal/GDALUtils.hpp>
+#include <pdal/private/gdal/GDALUtils.hpp>
 
 #include "private/Point.hpp"
 #include "private/pnp/GridPnp.hpp"
