+++ io/GDALWriter.cpp
@@ -36,8 +36,8 @@
 #include <sstream>
 
 #include <pdal/EigenUtils.hpp>
-#include <pdal/GDALUtils.hpp>
 #include <pdal/PointView.hpp>
+#include <pdal/private/gdal/Raster.hpp>
 
 #include "private/GDALGrid.hpp"
 
@@ -152,8 +152,6 @@ void GDALWriter::initialize()
     // don't expand by point if we're running in standard mode.  That's
     // set later in writeView.
     m_expandByPoint = !m_fixedGrid;
-
-    gdal::registerDrivers();
 }
