+++ pdal/Geometry.cpp
@@ -32,8 +32,12 @@
 * OF SUCH DAMAGE.
 ****************************************************************************/
 
+#include <ogr_api.h>
+#include <ogr_geometry.h>
+
 #include <pdal/Geometry.hpp>
-#include <pdal/GDALUtils.hpp>
+#include <pdal/private/gdal/GDALUtils.hpp>
+
 #include "private/SrsTransform.hpp"
 
 namespace pdal
