+++ pdal/Polygon.cpp
@@ -32,8 +32,11 @@
 * OF SUCH DAMAGE.
 ****************************************************************************/
 
-#include <pdal/GDALUtils.hpp>
+#include <ogr_api.h>
+#include <ogr_geometry.h>
+
 #include <pdal/Polygon.hpp>
+#include <pdal/private/gdal/GDALUtils.hpp>
 
 #include "../filters/private/pnp/GridPnp.hpp"
