+++ test/unit/filters/CropFilterTest.cpp
@@ -46,10 +46,6 @@
 #include <filters/StreamCallbackFilter.hpp>
 #include "Support.hpp"
 
-//ABELL
-#include <pdal/util/Bounds.hpp>
-#include <pdal/GDALUtils.hpp>
-
 using namespace pdal;
 
 TEST(CropFilterTest, create)
