+++ test/unit/io/GDALWriterTest.cpp
@@ -33,8 +33,8 @@
 ****************************************************************************/
 
 #include <pdal/pdal_test_main.hpp>
-#include <pdal/GDALUtils.hpp>
 #include <pdal/util/FileUtils.hpp>
+#include <pdal/private/gdal/Raster.hpp>
 #include <filters/RangeFilter.hpp>
 #include <io/BufferReader.hpp>
 #include <io/FauxReader.hpp>
@@ -99,7 +99,6 @@ void runGdalWriter(const Options& wo, const std::string& infile,
             arr.push_back(d);
         }
 
-        registerDrivers();
         Raster raster(outfile, "GTiff");
         if (raster.open() != GDALError::None)
         {
@@ -176,7 +175,6 @@ void runGdalWriter2(const Options& wo, const std::string& outfile,
         arr.push_back(d);
     }
 
-    registerDrivers();
     Raster raster(outfile, "GTiff");
     if (raster.open() != GDALError::None)
     {
@@ -581,7 +579,6 @@ TEST(GDALWriterTest, btint)
         arr.push_back(d);
     }
 
-    registerDrivers();
     Raster raster(outfile, "BT");
     if (raster.open() != GDALError::None)
     {
@@ -697,7 +694,6 @@ TEST(GDALWriterTest, issue_2074)
     w.prepare(t);
     w.execute(t);
 
-    gdal::registerDrivers();
     gdal::Raster raster(Support::temppath("gdal1.tif"), "GTiff");
     if (raster.open() != gdal::GDALError::None)
     {
@@ -860,7 +856,6 @@ TEST(GDALWriterTest, alternate_grid)
         w.prepare(t);
         w.execute(t);
 
-        gdal::registerDrivers();
         gdal::Raster raster(outfile, "GTiff");
         raster.open();
         EXPECT_EQ(raster.width(), 10);
